Refactored case list rows to be passed reason columns and lock tags
